Replace f64 with f32
WASM SIMD supports 128-bit registers, pack more variables in one

+++ src/camera.rs
@@ -4,25 +4,25 @@ use crate::{color, interval, object_list::object_list::ObjectList, ray::ray::Ray
 
 pub struct Camera {
     #[allow(dead_code)]
-    pub aspect_ratio: f64,
+    pub aspect_ratio: f32,
     pub image_width: u64,
     pub image_height: u64,
     pub samples_per_pixel: u64,
     pub max_depth: u64,
 
-    pub fov_vertical: f64,
-    pub location: Vector3<f64>,
-    pub look_at: Vector3<f64>,
-    pub up: Vector3<f64>,
+    pub fov_vertical: f32,
+    pub location: Vector3<f32>,
+    pub look_at: Vector3<f32>,
+    pub up: Vector3<f32>,
 
-    pixel_samples_scale:f64,
-    camera_centre: Vector3<f64>,
-    pixel_00_loc: Vector3<f64>,
-    pixel_delta_u: Vector3<f64>, 
-    pixel_delta_v: Vector3<f64>,
-    u: Vector3<f64>,
-    v: Vector3<f64>,
-    w: Vector3<f64>,    
+    pixel_samples_scale:f32,
+    camera_centre: Vector3<f32>,
+    pixel_00_loc: Vector3<f32>,
+    pixel_delta_u: Vector3<f32>, 
+    pixel_delta_v: Vector3<f32>,
+    u: Vector3<f32>,
+    v: Vector3<f32>,
+    w: Vector3<f32>,    
     texture: Vec<u8>, 
 }
 
@@ -30,7 +30,7 @@ impl Camera {
     pub fn set_resolution(&mut self, width: u64, height: u64) {
         self.image_width = width;
         self.image_height = height;
-        self.aspect_ratio = width as f64 / height as f64;
+        self.aspect_ratio = width as f32 / height as f32;
     }
     
     pub fn render(&mut self, world: ObjectList) {
@@ -57,8 +57,8 @@ impl Camera {
 
     fn get_ray(&self, i:u64, j:u64) -> Ray {
         let offset = Self::sample_square();
-        let pixel_sample = self.pixel_00_loc + ((i as f64 + offset.x) * self.pixel_delta_u)
-                            + ((j as f64 + offset.y) * self.pixel_delta_v);
+        let pixel_sample = self.pixel_00_loc + ((i as f32 + offset.x) * self.pixel_delta_u)
+                            + ((j as f32 + offset.y) * self.pixel_delta_v);
 
         let ray_origin = self.camera_centre;
         let ray_direction = pixel_sample - ray_origin;
@@ -66,19 +66,19 @@ impl Camera {
         return Ray::new(ray_origin, ray_direction);
     }
 
-    fn sample_square() -> Vector3<f64> {
+    fn sample_square() -> Vector3<f32> {
         // return Vector3::new(0.5, 0.5, 0.0);
-        return Vector3::new(fastrand::f64() - 0.5, fastrand::f64() - 0.5, 0.0);
+        return Vector3::new(fastrand::f32() - 0.5, fastrand::f32() - 0.5, 0.0);
     }
 
     fn initialise(&mut self) {
         self.camera_centre = self.location;
         let focal_length = (self.location - self.look_at).norm();
         let theta = self.fov_vertical.to_radians();
-        let h = f64::tan(theta/2.0);
+        let h = f32::tan(theta/2.0);
 
         let viewport_height = 2.0 * h * focal_length;
-        let viewport_width = viewport_height * ((self.image_width as f64)/self.image_height as f64);
+        let viewport_width = viewport_height * ((self.image_width as f32)/self.image_height as f32);
 
         self.w = (self.location - self.look_at).normalize();
         self.u = (self.up.cross(&self.w)).normalize();
@@ -87,19 +87,19 @@ impl Camera {
         let viewport_u = viewport_width * self.u;
         let viewport_v = viewport_height * -self.v;
         
-        self.pixel_delta_u = viewport_u.component_div(&Vector3::from_element(self.image_width as f64));
-        self.pixel_delta_v = viewport_v.component_div(&Vector3::from_element(self.image_height as f64));
+        self.pixel_delta_u = viewport_u.component_div(&Vector3::from_element(self.image_width as f32));
+        self.pixel_delta_v = viewport_v.component_div(&Vector3::from_element(self.image_height as f32));
 
         let viewport_upper_left = self.camera_centre - (focal_length * self.w) - viewport_u/2.0 - viewport_v/2.0; 
         self.pixel_00_loc = viewport_upper_left + 0.5 * (self.pixel_delta_u + self.pixel_delta_v);        
     }
 
-    fn ray_color(ray: &Ray, world: &ObjectList, depth: u64) -> Vector3<f64> {
+    fn ray_color(ray: &Ray, world: &ObjectList, depth: u64) -> Vector3<f32> {
         if depth <= 0 {
             return Vector3::new(0.0, 0.0, 0.0);
         }
 
-        if let Some(hit) = world.hit(ray, interval::Interval::new(0.001, std::f64::INFINITY)) {
+        if let Some(hit) = world.hit(ray, interval::Interval::new(0.001, std::f32::INFINITY)) {
             // let direction = hit.normal + vector_utils::random_vec3_unit();
             // return 0.5 * Self::ray_color(&Ray::new(hit.point, direction), world, depth - 1);
             let mut scattered: Ray = Ray::default();
@@ -120,9 +120,9 @@ impl Default for Camera {
     fn default() -> Self {
         let aspect_ratio = 16.0/9.0;
         let image_height = 1440;
-        let image_width = (aspect_ratio * image_height as f64) as u64;
+        let image_width = (aspect_ratio * image_height as f32) as u64;
         let samples_per_pixel = 4;
-        let pixel_samples_scale = 1.0 / samples_per_pixel as f64;
+        let pixel_samples_scale = 1.0 / samples_per_pixel as f32;
 
         return Camera{
             samples_per_pixel,
