+++ src/scene_object.rs
@@ -7,15 +7,15 @@ pub mod scene_object {
 
     #[derive(Clone)]
     pub struct HitRecord {
-        pub point: Vector3<f64>,
-        pub normal: Vector3<f64>,
+        pub point: Vector3<f32>,
+        pub normal: Vector3<f32>,
         pub material: Arc<dyn Material>,
-        pub t: f64,
+        pub t: f32,
         pub front_face: bool,
     }
 
     impl HitRecord {
-        /*pub fn set_face_normal(&mut self, ray: &ray::Ray, out_normal: &Vector3<f64>) {
+        /*pub fn set_face_normal(&mut self, ray: &ray::Ray, out_normal: &Vector3<f32>) {
             self.front_face = nalgebra::Vector::dot(&ray.direction(), &out_normal) < 0.0;
             
             if self.front_face {
